Type check for lang in config
The json library saves nulls as a special value, and not as null. Which makes sense and finally explains why the problem was there in the first place.
Also IsNullOrEmpty doesn't like anything that isn't a String. Which makes sense, to be honest

There's a weird thing that happen if the user customizes the separator, where it just resets after loading UTL once again. I've quickly looked into it, but I still haven't identified the cause